[MNG-8670] Test case for missing/wrong project events
Rather than adding another test, I've expanded on the one I previously
wrote for ProjectStarted events.
