+++ its/core-it-suite/src/test/resources/mng-8648/project/pom.xml
@@ -11,5 +11,7 @@
   <modules>
     <module>subproject-a</module>
     <module>subproject-b</module>
+    <module>subproject-c</module>
+    <module>subproject-d</module>
   </modules>
 </project>
