+++ its/core-it-suite/src/test/resources/mng-8648/project/subproject-c/pom.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/maven-v4_0_0.xsd">
+
+  <modelVersion>4.0.0</modelVersion>
+
+  <parent>
+    <groupId>org.apache.maven.its.mng8648</groupId>
+    <artifactId>root</artifactId>
+    <version>1-SNAPSHOT</version>
+  </parent>
+  <artifactId>subproject-c</artifactId>
+</project>
